+++ plugin-dev/Source/Sentry/Private/Mac/MacSentrySubsystem.cpp
@@ -11,12 +11,6 @@
 
 #include "GenericPlatform/GenericPlatformOutputDevices.h"
 
-#ifdef USE_SENTRY_SESSION_REPLAY
-#include "GenericPlatform/GenericPlatformSentryAttachment.h"
-#include "HAL/FileManager.h"
-#include "Misc/Guid.h"
-#endif
-
 void FMacSentrySubsystem::InitWithSettings(const USentrySettings* Settings, const FSentryCallbackHandlers& CallbackHandlers)
 {
 	FGenericPlatformSentrySubsystem::InitWithSettings(Settings, CallbackHandlers);
@@ -30,58 +24,7 @@ void FMacSentrySubsystem::InitWithSettings(const USentrySettings* Settings, cons
 	{
 		InitCrashReporter(Settings->GetEffectiveRelease(), Settings->GetEffectiveEnvironment());
 	}
-
-#ifdef USE_SENTRY_SESSION_REPLAY
-	if (IsEnabled() && Settings->AttachSessionReplay)
-	{
-		// Clear replay videos captured during previous session if any
-		IFileManager::Get().DeleteDirectory(*FPaths::Combine(GetDatabasePath(), TEXT("replays")), false, true);
-
-		SessionReplay = MakeUnique<FSentrySessionReplayRecorder>();
-		if (!SessionReplay->Initialize(Settings, GetReplayPath()))
-		{
-			SessionReplay.Reset();
-		}
-	}
-#endif
 }
-
-sentry_value_t FMacSentrySubsystem::OnCrash(const sentry_ucontext_t* uctx, sentry_value_t event, void* closure)
-{
-#ifdef USE_SENTRY_SESSION_REPLAY
-	if (SessionReplay && SessionReplay->HasSnapshotOnDisk())
-	{
-		TSharedPtr<ISentryAttachment> ReplayAttachment =
-			MakeShareable(new FGenericPlatformSentryAttachment(SessionReplay->GetAttachmentPath(), TEXT("session-replay.mp4"), TEXT("video/mp4")));
-
-		AddFileAttachment(ReplayAttachment);
-	}
-#endif
-
-	return FGenericPlatformSentrySubsystem::OnCrash(uctx, event, closure);
-}
-
-void FMacSentrySubsystem::Close()
-{
-#ifdef USE_SENTRY_SESSION_REPLAY
-	if (SessionReplay)
-	{
-		SessionReplay->Shutdown();
-		SessionReplay.Reset();
-	}
-#endif
-
-	FGenericPlatformSentrySubsystem::Close();
-}
-
-#ifdef USE_SENTRY_SESSION_REPLAY
-FString FMacSentrySubsystem::GetReplayPath() const
-{
-	const FString ReplayId = FGuid::NewGuid().ToString(EGuidFormats::DigitsWithHyphens).ToLower();
-	const FString ReplayPath = FPaths::Combine(GetDatabasePath(), TEXT("replays"), FString::Printf(TEXT("replay-%s.mp4"), *ReplayId));
-	return FPaths::ConvertRelativePathToFull(ReplayPath);
-}
-#endif
 
 FString FMacSentrySubsystem::GetHandlerExecutableName() const
 {