+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Add experimental Session Replay capturing on Mac ([#1425](https://github.com/getsentry/sentry-unreal/pull/1425))
+- Add experimental session replay capturing on Linux ([#1428](https://github.com/getsentry/sentry-unreal/pull/1428))
 
 ### Dependencies